+++ common/src/main/java/net/mehvahdjukaar/polytone/mixins/ParticleEngineMixin.java
@@ -1,16 +1,14 @@
 package net.mehvahdjukaar.polytone.mixins;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.PolytoneRenderTypes;
-import net.mehvahdjukaar.polytone.block.TickSource;
-import net.minecraft.client.Camera;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.particle.Particle;
 import net.minecraft.client.particle.ParticleEngine;
-import net.minecraft.client.renderer.LightTexture;
-import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.server.packs.resources.PreparableReloadListener;
 import net.minecraft.server.packs.resources.ResourceManager;
@@ -52,8 +50,4 @@ public abstract class ParticleEngineMixin {
         return size * 10;
     }
 
-    @Inject(method = "render", at = @At("HEAD"))
-    public void onRenderLast(LightTexture lightTexture, Camera camera, float partialTick, CallbackInfo ci) {
-        PolytoneRenderTypes.cacheMatrices();
-    }
 }