+++ forge/src/main/java/net/mehvahdjukaar/polytone/mixins/forge/ParticleEngineMixin2.java
@@ -1,0 +1,34 @@
+package net.mehvahdjukaar.polytone.mixins.forge;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.polytone.PolytoneRenderTypes;
+import net.minecraft.client.Camera;
+import net.minecraft.client.particle.ParticleEngine;
+import net.minecraft.client.particle.ParticleRenderType;
+import net.minecraft.client.renderer.LightTexture;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.culling.Frustum;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.function.Predicate;
+
+
+@Mixin(ParticleEngine.class)
+public abstract class ParticleEngineMixin2 {
+
+    @WrapOperation(
+            remap = false,
+            method = "render(Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/MultiBufferSource$BufferSource;Lnet/minecraft/client/renderer/LightTexture;Lnet/minecraft/client/Camera;FLnet/minecraft/client/renderer/culling/Frustum;)V",
+            at = @At(value = "INVOKE",
+                    ordinal = 0,
+                    target = "Lcom/mojang/blaze3d/systems/RenderSystem;applyModelViewMatrix()V"))
+    public void polytone$cacheMatrices(Operation<Void> original) {
+        original.call();
+        PolytoneRenderTypes.cacheMatrices(); //forge v
+    }
+}